css/css-regions/parsing/flow-into-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Regions Module Level 1: parsing flow-into with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-regions/#propdef-flow-into">
<meta name="assert" content="flow-into supports only the grammar 'none | <ident> [element|content]?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flow-into", "0");
test_invalid_value("flow-into", '"string"');
test_invalid_value("flow-into", "none COLUMN");
test_invalid_value("flow-into", "COLUMN none");
test_invalid_value("flow-into", "element");
test_invalid_value("flow-into", "element COLUMN");
test_invalid_value("flow-into", "COLUMN element content");
</script>
</body>
</html>
